feat: switch to composable executor for Optimism
